simple_drf/accounts/api/serializers.py
from rest_framework import serializers

from simple_drf.accounts.models import User


class UserSerializer(serializers.ModelSerializer[User]):
    class Meta:
        model = User
        fields = ["id", "username", "email", "first_name", "last_name", "date_joined"]
        read_only_fields = ["id"]
