pyproject.toml
[project]
name = "simple_drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "django>=5.1.1",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "django-allauth[socialaccount]>=64.2.1",
    "whitenoise[brotli]>=6.7.0",
    "djangorestframework>=3.15.2",
    "django-stubs[compatible-mypy]>=5.1.1",
    "djangorestframework-stubs[compatible-mypy]>=3.15.2",
]

[project.optional-dependencies]
prod = [
    "django-anymail[amazon-ses]>=12.0",
    "gunicorn>=23.0.0",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "django-coverage-plugin>=3.1.0",
    "factory-boy>=3.3.1",
]

[tool.djlint]
profile="django"
ignore="H021"
